Improved pylint rules complience
