+++ .pylintrc
@@ -1,9 +1,7 @@
 [MASTER]
 disable=
-    C0114, # missing-module-docstring
-    C0115,
-    C0116,
-    C0103,
-    W0613,
-    R0903,
-    E0239
+    missing-module-docstring,
+    missing-class-docstring,
+    missing-function-docstring,
+    unused-argument,
+    too-few-public-methods
