+++ src/api/common/file_helpers/upload_file_helper.py
@@ -5,7 +5,7 @@ from fastapi import UploadFile
 
 from yadisk import AsyncClient
 
-from src.api.common.file_helpers.BaseFileHelper import BaseFileHelper
+from src.api.common.file_helpers.base_file_helper import BaseFileHelper
 from src.api.common.enums import FileRetrievalCodes
 from src.app_config import app_config
 from src.utils import get_logger_from_filepath
