+++ src/api/common/file_helpers/ya_disk_helper.py
@@ -4,9 +4,9 @@ from pydantic import HttpUrl
 from yadisk import AsyncClient
 from yadisk.exceptions import NotFoundError
 
-from src.api.common.file_helpers.BaseFileHelper import BaseFileHelper
+from src.api.common.file_helpers.base_file_helper import BaseFileHelper
 from src.api.common.enums import FileRetrievalCodes
-from src.config.ConfigParser import ConfigParser
+from src.config.config_parser import ConfigParser
 
 
 class YaDiskHelper(BaseFileHelper):
