+++ src/api/common/queue_request.py
@@ -7,7 +7,7 @@ from starlette.responses import Response
 from src.api.common.types.request import RequestType
 from src.api.common.handlers import global_requests_handler
 from src.api.common.enums import RequestProcessCodes
-from src.api.common.schemas.MediaRequest import MediaRequestDTO, MediaRequestSchema
+from src.api.common.schemas.media_request import MediaRequestDTO, MediaRequestSchema
 
 
 async def queue_request(
