+++ src/api/common/request_helpers/__init__.py
@@ -1,3 +1,3 @@
-from .TranscriptionHelper import TranscriptionHelper
-from .GigachatHelper import GigachatHelper
-from .HelpersHandler import HelpersHandler
+from .transcription_helper import TranscriptionHelper
+from .gigachat_helper import GigachatHelper
+from .helpers_handler import HelpersHandler
