+++ src/api/common/request_helpers/gigachat_helper.py
@@ -4,9 +4,9 @@ from gigachat import GigaChat
 from gigachat.models import Messages, MessagesRole, Chat
 
 from src.api.common.enums import RequestHelpersNames
-from src.api.common.request_helpers.BaseHelper import BaseHelper
+from src.api.common.request_helpers.base_helper import BaseHelper
 from src.api.common.utils import get_summary_filename
-from src.api.video.schemas.requests.Summarize import SummarizeConfig
+from src.api.video.schemas.requests.summarize import SummarizeConfig
 from src.app_config import app_config
 from src.config.enums import GigachatModels
 from src.config.schemas.default_system_prompt import SYSTEM_PROMPT
