+++ src/api/common/schemas/__init__.py
@@ -1,1 +1,1 @@
-from .MediaRequest import MediaRequestSchema
+from .media_request import MediaRequestSchema
