+++ src/api/common/services/request_queue.py
@@ -1,7 +1,7 @@
 import asyncio
 
 from src.api.common.types.request import RequestType
-from src.api.common.schemas.MediaRequest import MediaRequestDTO
+from src.api.common.schemas.media_request import MediaRequestDTO
 from src.api.video.constants import MAX_REQUESTS_BACKLOG
