+++ src/api/common/types/request_helper.py
@@ -1,3 +1,3 @@
-from src.api.common.request_helpers import BaseHelper
+from src.api.common.request_helpers.base_helper import BaseHelper
 
 RequestHelper = BaseHelper
