+++ src/api/video/schemas/requests/__init__.py
@@ -1,5 +1,5 @@
-from .Compress import CompressSchema
-from .Transcribe import TranscribeSchema
-from .Summarize import SummarizeSchema
-from .ExtractAudio import ExtractAudioSchema
-from .Custom import CustomSchema
+from .compress import CompressSchema
+from .transcribe import TranscribeSchema
+from .summarize import SummarizeSchema
+from .extract_audio import ExtractAudioSchema
+from .custom import CustomSchema
