+++ src/api/video/schemas/requests/custom.py
@@ -2,10 +2,10 @@ from pydantic import Field
 
 from src.api.common.schemas import MediaRequestSchema
 from src.api.video.enums import VideoRequestType
-from src.api.video.schemas.requests.Compress import CompressConfig
-from src.api.video.schemas.requests.ExtractAudio import AudioSettings
-from src.api.video.schemas.requests.Summarize import SummarizeConfig
-from src.api.video.schemas.requests.Transcribe import TranscribeConfig
+from src.api.video.schemas.requests.compress import CompressConfig
+from src.api.video.schemas.requests.extract_audio import AudioSettings
+from src.api.video.schemas.requests.summarize import SummarizeConfig
+from src.api.video.schemas.requests.transcribe import TranscribeConfig
 
 
 class CustomConfig(AudioSettings, SummarizeConfig, TranscribeConfig):
