+++ src/api/video/schemas/requests/summarize.py
@@ -3,7 +3,7 @@ from pydantic import Field
 from src.api.common.schemas import MediaRequestSchema
 from src.app_config import app_config
 from src.config.enums import GigachatModels
-from src.config.schemas.BaseEnumModel import BaseEnumModel
+from src.config.schemas.base_enum_model import BaseEnumModel
 
 
 class SummarizeSettings(BaseEnumModel):
