+++ src/api/video/services/__init__.py
@@ -1,5 +1,5 @@
-from .Summarize import SummarizeHandler as Summarization
-from .ExtractAudio import ExtractAudioHandler as AudioExtraction
-from .Transcribe import TranscriptionHandler as Transcription
-from .Compress import CompressHandler as Compression
-from .Custom import CustomHandler as Custom
+from .summarize import SummarizeHandler as Summarization
+from .extract_audio import ExtractAudioHandler as AudioExtraction
+from .transcribe import TranscriptionHandler as Transcription
+from .compress import CompressHandler as Compression
+from .custom import CustomHandler as Custom
