+++ src/api/video/services/compress.py
@@ -1,15 +1,15 @@
 from pathlib import Path
 
-from src.api.common.services.BaseHandler import BaseHandler
 from src.api.common.types.request import RequestType
+from src.pipeline.render import RendererBuilder, Renderer
+from src.api.common.services.base_handler import BaseHandler
 from src.api.video.enums import VideoRequestType
-from src.pipeline.render import Renderer, RendererBuilder
 from src.pipeline.tasks import jobs
 
 
-class ExtractAudioHandler(BaseHandler):
+class CompressHandler(BaseHandler):
     def __init__(self):
-        super().__init__(VideoRequestType.EXTRACT_AUDIO)
+        super().__init__(VideoRequestType.COMPRESS)
 
     def _build_renderer(
         self, actions: list[RequestType], raw_file_path: Path
@@ -17,5 +17,6 @@ class ExtractAudioHandler(BaseHandler):
         return (
             RendererBuilder()
             .use_file(str(raw_file_path))
-            .add_task(jobs.ExtractAudioTask())
-        ).build()
+            .add_task(jobs.TwoPassEncodingTask())
+            .build()
+        )
