+++ src/api/video/services/extract_audio.py
@@ -1,15 +1,15 @@
 from pathlib import Path
 
-from src.api.common.services.BaseHandler import BaseHandler
+from src.api.common.services.base_handler import BaseHandler
 from src.api.common.types.request import RequestType
 from src.api.video.enums import VideoRequestType
-from src.pipeline.render import RendererBuilder, Renderer
+from src.pipeline.render import Renderer, RendererBuilder
 from src.pipeline.tasks import jobs
 
 
-class TranscriptionHandler(BaseHandler):
+class ExtractAudioHandler(BaseHandler):
     def __init__(self):
-        super().__init__(VideoRequestType.TRANSCRIBE)
+        super().__init__(VideoRequestType.EXTRACT_AUDIO)
 
     def _build_renderer(
         self, actions: list[RequestType], raw_file_path: Path
@@ -17,5 +17,5 @@ class TranscriptionHandler(BaseHandler):
         return (
             RendererBuilder()
             .use_file(str(raw_file_path))
-            .add_task(jobs.TranscribeTask())
+            .add_task(jobs.ExtractAudioTask())
         ).build()
