+++ src/api/video/services/summarize.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from src.api.common.services.BaseHandler import BaseHandler
+from src.api.common.services.base_handler import BaseHandler
 from src.api.common.types.request import RequestType
 from src.api.video.enums import VideoRequestType
 from src.pipeline.render import Renderer, RendererBuilder
