+++ src/api/video/services/transcribe.py
@@ -1,15 +1,15 @@
 from pathlib import Path
 
+from src.api.common.services.base_handler import BaseHandler
 from src.api.common.types.request import RequestType
-from src.pipeline.render import RendererBuilder, Renderer
-from src.api.common.services.BaseHandler import BaseHandler
 from src.api.video.enums import VideoRequestType
+from src.pipeline.render import RendererBuilder, Renderer
 from src.pipeline.tasks import jobs
 
 
-class CompressHandler(BaseHandler):
+class TranscriptionHandler(BaseHandler):
     def __init__(self):
-        super().__init__(VideoRequestType.COMPRESS)
+        super().__init__(VideoRequestType.TRANSCRIBE)
 
     def _build_renderer(
         self, actions: list[RequestType], raw_file_path: Path
@@ -17,6 +17,5 @@ class CompressHandler(BaseHandler):
         return (
             RendererBuilder()
             .use_file(str(raw_file_path))
-            .add_task(jobs.TwoPassEncodingTask())
-            .build()
-        )
+            .add_task(jobs.TranscribeTask())
+        ).build()
