+++ src/app_config.py
@@ -1,4 +1,4 @@
-from src.config.ConfigParser import ConfigParser
+from src.config.config_parser import ConfigParser
 from src.constants import CONFIG_PATH
 
 app_config: ConfigParser = ConfigParser.from_yaml(CONFIG_PATH)
