+++ src/config/schemas/base_enum_model.py
