+++ src/config/schemas/codecs.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 
 from src.config.enums import AudioCodecs, VideoCodecs
-from src.config.schemas.BaseEnumModel import BaseEnumModel
+from src.config.schemas.base_enum_model import BaseEnumModel
 
 
 class CodecsSchema(BaseEnumModel):
