+++ src/config/schemas/summary.py
@@ -1,6 +1,6 @@
 from pydantic import Field
 from src.config.enums import GigachatModels, GigachatModelScopes
-from src.config.schemas.BaseEnumModel import BaseEnumModel
+from src.config.schemas.base_enum_model import BaseEnumModel
 
 
 class SummarySchema(BaseEnumModel):
