+++ src/config/schemas/video.py
@@ -1,6 +1,6 @@
 from pydantic import Field
 
-from src.config.schemas.BaseEnumModel import BaseEnumModel
+from src.config.schemas.base_enum_model import BaseEnumModel
 
 
 class VideoSchema(BaseEnumModel):
