+++ src/pipeline/schemas/paths.py
