+++ src/pipeline/tasks/jobs/extract_audio.py
@@ -3,10 +3,10 @@ import ffmpeg
 from src.api.common.request_helpers import HelpersHandler
 from src.api.common.types.request import RequestType
 from src.api.video.enums import VideoRequestType
-from src.api.video.schemas.requests.ExtractAudio import ExtractAudioConfig
-from src.pipeline.schemas.Paths import PathsSchema
-from src.pipeline.schemas.Streams import StreamsSchema
-from src.pipeline.tasks.jobs.BaseJob import BaseJob
+from src.api.video.schemas.requests.extract_audio import ExtractAudioConfig
+from src.pipeline.schemas.paths import PathsSchema
+from src.pipeline.schemas.streams import StreamsSchema
+from src.pipeline.tasks.jobs.base_job import BaseJob
 from src.pipeline.tasks.utils import extract_config_by_field_name
 from src.pipeline.types import RenderConfig
