+++ src/pipeline/tasks/jobs/transcribe.py
@@ -2,19 +2,19 @@ from src.api.common.enums import RequestHelpersNames
 from src.api.common.request_helpers import HelpersHandler, TranscriptionHelper
 from src.api.common.types.request import RequestType
 from src.api.video.enums import VideoRequestType
-from src.api.video.schemas.requests.Transcribe import TranscribeConfig
-from src.pipeline.BaseTask import BaseTask
-from src.pipeline.schemas.Paths import PathsSchema
-from src.pipeline.schemas.Streams import StreamsSchema
-from src.pipeline.tasks import jobs
-from src.pipeline.tasks.jobs.BaseJob import BaseJob
+from src.api.video.schemas.requests.transcribe import TranscribeConfig
+from src.pipeline.base_task import BaseTask
+from src.pipeline.schemas.paths import PathsSchema
+from src.pipeline.schemas.streams import StreamsSchema
+from src.pipeline.tasks.jobs.extract_audio import ExtractAudioTask
+from src.pipeline.tasks.jobs.base_job import BaseJob
 from src.pipeline.tasks.utils import extract_config_by_field_name
 from src.pipeline.types import RenderConfig
 
 
 class TranscribeTask(BaseJob):
     request_type: RequestType = VideoRequestType.TRANSCRIBE
-    dependencies: list[BaseTask] = [jobs.ExtractAudioTask()]
+    dependencies: list[BaseTask] = [ExtractAudioTask()]
 
     @staticmethod
     def extract_config(full_config: RenderConfig) -> TranscribeConfig:
