+++ src/pipeline/types/preprocessor.py
@@ -1,8 +1,8 @@
 from typing import Callable, Awaitable
 
 from src.api.common.request_helpers import HelpersHandler
-from src.pipeline.schemas.Paths import PathsSchema
-from src.pipeline.schemas.Streams import StreamsSchema
+from src.pipeline.schemas.paths import PathsSchema
+from src.pipeline.schemas.streams import StreamsSchema
 from src.pipeline.types.render_config import RenderConfig
 
 Preprocessor = Callable[
