+++ src/pipeline/types/render_config.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
 
-from src.config.schemas.BaseEnumModel import BaseEnumModel
+from src.config.schemas.base_enum_model import BaseEnumModel
 
 RenderConfig = BaseModel | BaseEnumModel
